develoepd patent extraction code from Scopus
1) renamed file with patent results
2) adding acknowledgements into universities_table.csv
3) tested the code to extraction patent amount from Scopus. Here,
possible problems in extracting the amount of patents for each
affiliation might be that the I could not code selenium to wait until
the page is fully loaded after the Keys.RETURN key is pressed to send
the query to the scopus web-site. In order to mitigate that sime hard
time wait is added (time.sleep() function). Hopefully, this will work
without problem.

TODO:
1) add driver.find_element_by_id(...) with a class of hidden to ensure
that the affiliation did not publish any patent at that year (the button
with patents is disabled)

 On branch master
 Changes to be committed:
	new file:   data/patent_count.csv
	deleted:    data/patent_count_2016.csv
	modified:   data/universities_table.csv
	modified:   src/get_patent_count.py